+++ packages/xchain-bitcoin/src/index.ts
@@ -1,4 +1,4 @@
 export * from './types'
 export * from './client'
 export * from './const'
-export { getDefaultFees, getDefaultFeesWithRates, getPrefix, validateAddress, calcFee } from './utils'
+export { getPrefix, validateAddress } from './utils'