+++ packages/xchain-doge/src/utils.ts
@@ -1,57 +1,16 @@
-import { FeeRate, Fees, FeesWithRates, Network, UTXO, calcFees, standardFeeRates } from '@xchainjs/xchain-client'
-import { Address, BaseAmount, baseAmount } from '@xchainjs/xchain-util'
+import { Network } from '@xchainjs/xchain-client'
+import { Address } from '@xchainjs/xchain-util'
 import * as Dogecoin from 'bitcoinjs-lib'
 import coininfo from 'coininfo'
 
-import { MIN_TX_FEE } from './const'
+export const TX_EMPTY_SIZE = 4 + 1 + 1 + 4 //10
+export const TX_INPUT_BASE = 32 + 4 + 1 + 4 // 41
+export const TX_INPUT_PUBKEYHASH = 107
+export const TX_OUTPUT_BASE = 8 + 1 //9
+export const TX_OUTPUT_PUBKEYHASH = 25
 
-const TX_EMPTY_SIZE = 4 + 1 + 1 + 4 //10
-const TX_INPUT_BASE = 32 + 4 + 1 + 4 // 41
-const TX_INPUT_PUBKEYHASH = 107
-const TX_OUTPUT_BASE = 8 + 1 //9
-const TX_OUTPUT_PUBKEYHASH = 25
-
-function inputBytes(): number {
+export function inputBytes(): number {
   return TX_INPUT_BASE + TX_INPUT_PUBKEYHASH
-}
-
-/**
- * Compile memo.
- *
- * @param {string} memo The memo to be compiled.
- * @returns {Buffer} The compiled memo.
- */
-export const compileMemo = (memo: string): Buffer => {
-  const data = Buffer.from(memo, 'utf8') // converts MEMO to buffer
-  return Dogecoin.script.compile([Dogecoin.opcodes.OP_RETURN, data]) // Compile OP_RETURN script
-}
-
-/**
- * Get the transaction fee.
- *
- * @param {UTXO[]} inputs The UTXOs.
- * @param {FeeRate} feeRate The fee rate.
- * @param {Buffer} data The compiled memo (Optional).
- * @returns {number} The fee amount.
- */
-export function getFee(inputs: UTXO[], feeRate: FeeRate, data: Buffer | null = null): number {
-  const inputSizeBasedOnInputs =
-    inputs.length > 0
-      ? inputs.reduce((a) => a + inputBytes(), 0) + inputs.length // +1 byte for each input signature
-      : (TX_INPUT_BASE + TX_INPUT_PUBKEYHASH) * 2 + 2 // By default 2 UTXOs // Temporal solution until issue addressed https://github.com/xchainjs/xchainjs-lib/issues/850
-  let sum =
-    TX_EMPTY_SIZE +
-    inputSizeBasedOnInputs + // +1 byte for each input signature
-    TX_OUTPUT_BASE +
-    TX_OUTPUT_PUBKEYHASH +
-    TX_OUTPUT_BASE +
-    TX_OUTPUT_PUBKEYHASH
-
-  if (data) {
-    sum += TX_OUTPUT_BASE + data.length
-  }
-  const fee = sum * feeRate
-  return fee > MIN_TX_FEE ? fee : MIN_TX_FEE
 }
 
 /**
@@ -105,45 +64,6 @@ export const validateAddress = (address: Address, network: Network): boolean => 
   } catch (error) {
     return false
   }
-}
-
-/**
- * Calculate fees based on fee rate and memo.
- *
- * @param {FeeRate} feeRate
- * @param {string} memo
- * @returns {BaseAmount} The calculated fees based on fee rate and the memo.
- */
-export const calcFee = (feeRate: FeeRate, memo?: string): BaseAmount => {
-  const compiledMemo = memo ? compileMemo(memo) : null
-  const fee = getFee([], feeRate, compiledMemo)
-  return baseAmount(fee)
-}
-
-/**
- * Get the default fees with rates.
- *
- * @returns {FeesWithRates} The default fees and rates.
- */
-export const getDefaultFeesWithRates = (): FeesWithRates => {
-  const rates = {
-    ...standardFeeRates(MIN_TX_FEE),
-  }
-
-  return {
-    fees: calcFees(rates, calcFee),
-    rates,
-  }
-}
-
-/**
- * Get the default fees.
- *
- * @returns {Fees} The default fees.
- */
-export const getDefaultFees = (): Fees => {
-  const { fees } = getDefaultFeesWithRates()
-  return fees
 }
 
 /**