+++ packages/xchain-litecoin/src/client.ts
@@ -1,6 +1,5 @@
 import {
   AssetInfo,
-  Fee,
   FeeOption,
   FeeRate,
   Network,
@@ -23,6 +22,7 @@ import {
   LOWER_FEE_BOUND,
   LTCChain,
   LTC_DECIMAL,
+  MIN_TX_FEE,
   UPPER_FEE_BOUND,
   explorerProviders,
 } from './const'
@@ -171,10 +171,6 @@ class Client extends UTXOClient {
     }
   }
 
-  protected async calcFee(feeRate: FeeRate, memo?: string): Promise<Fee> {
-    return Utils.calcFee(feeRate, memo)
-  }
-
   /**
    * Transfer LTC.
    *
@@ -204,7 +200,8 @@ class Client extends UTXOClient {
       auth: this.nodeAuth,
     })
   }
-  buildTx = async ({
+
+  async buildTx({
     amount,
     recipient,
     memo,
@@ -213,14 +210,14 @@ class Client extends UTXOClient {
   }: TxParams & {
     feeRate: FeeRate
     sender: Address
-  }): Promise<{ psbt: Litecoin.Psbt; utxos: UTXO[] }> => {
+  }): Promise<{ psbt: Litecoin.Psbt; utxos: UTXO[] }> {
     if (!this.validateAddress(recipient)) throw new Error('Invalid address')
 
     const utxos = await this.scanUTXOs(sender, false)
     if (utxos.length === 0) throw new Error('No utxos to send')
 
     const feeRateWhole = Number(feeRate.toFixed(0))
-    const compiledMemo = memo ? Utils.compileMemo(memo) : null
+    const compiledMemo = memo ? this.compileMemo(memo) : null
 
     const targetOutputs = []
 
@@ -266,6 +263,46 @@ class Client extends UTXOClient {
     })
 
     return { psbt, utxos }
+  }
+
+  /**
+   * Compile memo.
+   *
+   * @param {string} memo The memo to be compiled.
+   * @returns {Buffer} The compiled memo.
+   */
+  protected compileMemo(memo: string): Buffer {
+    const data = Buffer.from(memo, 'utf8') // converts MEMO to buffer
+    return Litecoin.script.compile([Litecoin.opcodes.OP_RETURN, data]) // Compile OP_RETURN script
+  }
+
+  /**
+   * Get the transaction fee.
+   *
+   * @param {UTXO[]} inputs The UTXOs.
+   * @param {FeeRate} feeRate The fee rate.
+   * @param {Buffer} data The compiled memo (Optional).
+   * @returns {number} The fee amount.
+   */
+  protected getFeeFromUtxos(inputs: UTXO[], feeRate: FeeRate, data: Buffer | null = null): number {
+    const inputSizeBasedOnInputs =
+      inputs.length > 0
+        ? inputs.reduce((a, x) => a + Utils.inputBytes(x), 0) + inputs.length // +1 byte for each input signature
+        : 0
+    let sum =
+      Utils.TX_EMPTY_SIZE +
+      inputSizeBasedOnInputs +
+      inputs.length + // +1 byte for each input signature
+      Utils.TX_OUTPUT_BASE +
+      Utils.TX_OUTPUT_PUBKEYHASH +
+      Utils.TX_OUTPUT_BASE +
+      Utils.TX_OUTPUT_PUBKEYHASH
+
+    if (data) {
+      sum += Utils.TX_OUTPUT_BASE + data.length
+    }
+    const fee = sum * feeRate
+    return fee > MIN_TX_FEE ? fee : MIN_TX_FEE
   }
 }