+++ packages/xchain-litecoin/src/types/common.ts
@@ -2,13 +2,6 @@ export type Witness = {
   value: number
   script: Buffer
 }
-export type UTXO = {
-  hash: string
-  index: number
-  value: number
-  witnessUtxo: Witness
-  txHex?: string
-}
 
 export type NodeAuth = {
   username: string